refactor: extract event viewer
