+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { EventViewer } from "./EventViewer";
+export { NavigationBar } from "./NavigationBar";
